# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma Fintech de Billeteras Digitales API
  version: 1.0.0
  description: |
    Contrato REST entre el backend Spring Boot y el frontend React.
    Este archivo es la fuente de verdad para DTOs, endpoints y errores.
servers:
  - url: http://localhost:8080/api/v1
    description: Backend local Spring Boot
...
paths:
...
  /users/{userId}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Obtener usuario por ID
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      operationId: updateUser
      summary: Actualizar nombre o email del usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Eliminar usuario y datos en cascada
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Usuario eliminado
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}/wallets:
    get:
      tags: [Wallets]
      operationId: getUserWallets
      summary: Listar billeteras de un usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Billeteras del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalletResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Wallets]
      operationId: createWallet
      summary: Crear billetera para usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Billetera creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{userId}/wallets/{walletId}/recharge:
    post:
      tags: [Operations]
      operationId: rechargeWallet
      summary: Recargar billetera
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/WalletId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoneyOperationRequest'
      responses:
        '200':
          description: Recarga realizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessError'
  /users/{userId}/wallets/{walletId}/withdraw:
    post:
      tags: [Operations]
      operationId: withdrawWallet
      summary: Retirar dinero de billetera
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/WalletId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoneyOperationRequest'
      responses:
        '200':
          description: Retiro realizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessError'
  /users/{userId}/transfers/internal:
    post:
      tags: [Operations]
      operationId: createInternalTransfer
      summary: Transferir entre billeteras del mismo usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalTransferRequest'
      responses:
        '200':
          description: Transferencia interna realizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessError'
...
  /users/{userId}/points:
    get:
      tags: [Points]
      operationId: getUserPoints
      summary: Obtener puntos y nivel del usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Puntos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /points/ranking:
    get:
      tags: [Points]
      operationId: getPointsRanking
      summary: Obtener ranking de puntos
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Ranking descendente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingItemResponse'
...
  /notifications/users/{userId}:
    get:
      tags: [Notifications]
      operationId: getUserNotifications
      summary: Listar notificaciones del usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notificaciones del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationAsRead
      summary: Marcar notificación como leída
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: Notificación actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /fraud/events:
    get:
      tags: [Fraud]
      operationId: getFraudEvents
      summary: Listar eventos sospechosos
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
        - name: severity
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FraudSeverity'
      responses:
        '200':
          description: Eventos sospechosos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FraudEventResponse'
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      example: USR001
    WalletId:
      name: walletId
      in: path
      required: true
      schema:
        type: string
      example: WALLET-001
...
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
      example: NOTIF-001
...
  responses:
    BadRequest:
      description: Request inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Conflicto de recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BusinessError:
      description: Regla de negocio incumplida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    UserResponse:
      type: object
      required: [id, name, email, registeredAt, points, loyaltyLevel, walletCount, totalBalance]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        registeredAt:
          type: string
          format: date-time
        points:
          type: number
          format: double
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
        walletCount:
          type: integer
        totalBalance:
          type: number
          format: double
    CreateWalletRequest:
      type: object
      required: [code, name, type]
      properties:
        code:
          type: string
          example: WALLET-001
        name:
          type: string
          example: Gastos diarios
        type:
          type: string
          example: DAILY_EXPENSES
    WalletResponse:
      type: object
      required: [code, name, type, ownerId, balance, active, createdAt, transactionCount]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        ownerId:
          type: string
        balance:
          type: number
          format: double
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        transactionCount:
          type: integer
    MoneyOperationRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 1500.0
        description:
          type: string
          example: Recarga inicial
    InternalTransferRequest:
      type: object
      required: [sourceWalletId, targetWalletId, amount]
      properties:
        sourceWalletId:
          type: string
        targetWalletId:
          type: string
        amount:
          type: number
          format: double
          minimum: 0.01
        description:
          type: string
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Juan Actualizado
        email:
          type: string
          format: email
          example: actualizado@example.com
...
    TransactionResponse:
      type: object
      required:
        - id
        - timestamp
        - type
        - amount
        - sourceWalletId
        - targetWalletId
        - sourceUserId
        - targetUserId
        - status
        - pointsGenerated
        - reversible
        - riskLevel
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: double
        sourceWalletId:
          type: string
        targetWalletId:
          type: string
        sourceUserId:
          type: string
        targetUserId:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        pointsGenerated:
          type: number
          format: double
        description:
          type: string
        reversible:
          type: boolean
        riskLevel:
          $ref: '#/components/schemas/FraudSeverity'
    PointsResponse:
      type: object
      required: [userId, points, loyaltyLevel]
      properties:
        userId:
          type: string
        points:
          type: number
          format: double
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
    RankingItemResponse:
      type: object
      required: [position, userId, userName, points, loyaltyLevel]
      properties:
        position:
          type: integer
        userId:
          type: string
        userName:
          type: string
        points:
          type: number
          format: double
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
...
    NotificationResponse:
      type: object
      required: [id, userId, type, severity, title, message, read, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: '#/components/schemas/NotificationType'
        severity:
          $ref: '#/components/schemas/NotificationSeverity'
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    FraudEventResponse:
      type: object
      required: [id, userId, type, severity, description, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        transactionId:
          type: string
        type:
          type: string
          example: HIGH_VELOCITY
        severity:
          $ref: '#/components/schemas/FraudSeverity'
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: El request es inválido.
        details:
          type: array
          items:
            type: string
    LoyaltyLevel:
      type: string
      enum: [BRONZE, SILVER, GOLD, PLATINUM]
    TransactionType:
      type: string
      enum: [RECHARGE, WITHDRAWAL, INTERNAL_TRANSFER, EXTERNAL_TRANSFER_SENT, EXTERNAL_TRANSFER_RECEIVED]
    TransactionStatus:
      type: string
      enum: [SUCCESSFUL, REVERSED, PENDING, REJECTED]
...
    NotificationType:
      type: string
      enum: [LOW_BALANCE, TRANSACTION, FRAUD_ALERT, POINTS_LEVEL, SYSTEM]
    NotificationSeverity:
      type: string
      enum: [INFO, WARNING, CRITICAL]
    FraudSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]